pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snowl"
version = "0.1.0"
description = "Snowl: a general agent evaluation framework."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.25",
  "PyYAML>=6.0",
  "requests>=2.31",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "build>=1.0.0",
  "twine>=5.0.0",
]
web = [
  "fastapi>=0.111",
  "uvicorn>=0.30",
]
osworld_eval = [
  "beautifulsoup4",
  "borb",
  "cssselect",
  "easyocr",
  "imagehash",
  "librosa",
  "lxml",
  "mutagen",
  "numpy",
  "odfpy",
  "opencv-python-headless",
  "openpyxl",
  "pandas",
  "pdfplumber",
  "pillow",
  "playwright",
  "pydrive",
  "pymupdf",
  "pypdf",
  "pypdf2",
  "python-docx",
  "python-pptx",
  "pyacoustid",
  "pytz",
  "rapidfuzz",
  "requests-toolbelt",
  "scikit-image",
  "tldextract>=5.3.0",
  "xmltodict",
]

[project.scripts]
snowl = "snowl.cli:main"

[project.urls]
Homepage = "https://pypi.org/project/snowl/"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["snowl", "snowl.*"]

[tool.setuptools.package-data]
"snowl.ui" = ["panel_configs/*.yml", "panel_configs/*.yaml", "panel_configs/README.md"]
